+++ services/121-service/src/config.ts
@@ -57,4 +57,6 @@ export const EXTERNAL_API = {
     process.env.EXTERNAL_121_SERVICE_URL + 'api/notifications/voice/xml/',
   voiceMp3lUrl:
     process.env.EXTERNAL_121_SERVICE_URL + 'api/notifications/voice/mp3/',
+  africasTalkingValidationUrl:
+    process.env.EXTERNAL_121_SERVICE_URL + 'api/fsp/africastalking/validation/',
 };