+++ services/121-service/src/shared/dto/status-message.dto.ts
@@ -1,0 +1,6 @@
+import { StatusEnum } from "../enum/status.enum";
+
+export class StatusMessageDto {
+  public readonly status: StatusEnum;
+  public readonly message: any;
+}